requirements.txt
# Django e REST Framework
Django==6.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
django-filter==24.2

# Banco de dados
psycopg2-binary==2.9.9

# Variáveis de ambiente
python-dotenv==1.2.1

# Qualidade de código e testes
pytest==8.3.3
pytest-django==4.8.0
pytest-cov==5.0.0
coverage==7.5.3
black==24.8.0
flake8==7.1.1